+++ app/cli/cli.py
@@ -5,6 +5,7 @@ from mediagrabber.core import MediaGrabber
 import logging
 from mediagrabber.config import Config
 
+
 def main():
     # Set desired logging level
     logging.basicConfig(level=Config.log_level())
@@ -14,6 +15,7 @@ def main():
 
     service: MediaGrabber = injector.get(MediaGrabber)
     fire.Fire(service)
+
 
 if __name__ == "__main__":
     main()
